# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PersonaFit API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/surveys": {
      "post": {
        "tags": [
          "surveys"
        ],
        "summary": "Create Survey",
        "operationId": "create_survey_api_surveys_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "surveys"
        ],
        "summary": "List Surveys Endpoint",
        "operationId": "list_surveys_endpoint_api_surveys_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "running",
                    "completed",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/surveys/{survey_id}": {
      "get": {
        "tags": [
          "surveys"
        ],
        "summary": "Get Survey Endpoint",
        "operationId": "get_survey_endpoint_api_surveys__survey_id__get",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Survey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "surveys"
        ],
        "summary": "Update Survey Endpoint",
        "operationId": "update_survey_endpoint_api_surveys__survey_id__put",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Survey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "surveys"
        ],
        "summary": "Delete Survey Endpoint",
        "operationId": "delete_survey_endpoint_api_surveys__survey_id__delete",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Survey Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/surveys/{survey_id}/report": {
      "get": {
        "tags": [
          "survey_report"
        ],
        "summary": "Get Report",
        "description": "완료된 응답 전체 집계. 백엔드에서 사전 가공해 프론트 부담 제거.",
        "operationId": "get_report_api_surveys__survey_id__report_get",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Survey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/surveys/{survey_id}/commentary/regenerate": {
      "post": {
        "tags": [
          "survey_report"
        ],
        "summary": "Regenerate Commentary",
        "description": "비동기 백필. 즉시 202 반환 후 background에서 LLM 호출 + 저장.",
        "operationId": "regenerate_commentary_api_surveys__survey_id__commentary_regenerate_post",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Survey Id"
            }
          },
          {
            "name": "X-Admin-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Regenerate Commentary Api Surveys  Survey Id  Commentary Regenerate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/surveys/{survey_id}/report.csv": {
      "get": {
        "tags": [
          "survey_report"
        ],
        "summary": "Get Report Csv",
        "description": "응답 raw 데이터 CSV — 1행 = 페르소나 1명.",
        "operationId": "get_report_csv_api_surveys__survey_id__report_csv_get",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Survey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutionConfig": {
        "properties": {
          "llm_provider": {
            "type": "string",
            "enum": [
              "anthropic",
              "sllm"
            ],
            "title": "Llm Provider",
            "default": "anthropic"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "claude-haiku-4-5"
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "include_reasoning": {
            "type": "boolean",
            "title": "Include Reasoning",
            "default": true
          }
        },
        "type": "object",
        "title": "ExecutionConfig",
        "description": "LLM 실행 파라미터."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Question": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "uuid4"
          },
          "order": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Order",
            "description": "1-based 노출 순서"
          },
          "type": {
            "type": "string",
            "enum": [
              "single_choice",
              "multi_choice",
              "scale",
              "open_ended",
              "nps"
            ],
            "title": "Type"
          },
          "text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Text"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Options"
          },
          "scale_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Min"
          },
          "scale_max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Max"
          },
          "scale_label_low": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Label Low"
          },
          "scale_label_high": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Label High"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "order",
          "type",
          "text"
        ],
        "title": "Question",
        "description": "설문 1문항.\n\n유형별 필드 사용:\n  - single_choice / multi_choice: options 필수 (최소 2개)\n  - scale: scale_min/scale_max 필수 (예: 1-5, 1-7)\n  - open_ended: 추가 필드 없음\n  - nps: scale_min=0, scale_max=10 고정"
      },
...
      "ReportResponse": {
        "properties": {
          "survey": {
            "$ref": "#/components/schemas/Survey"
          },
          "summary": {
            "$ref": "#/components/schemas/ReportSummary"
          },
          "respondent_distribution": {
            "$ref": "#/components/schemas/RespondentDistribution"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionReport"
            },
            "type": "array",
            "title": "Questions"
          },
          "overall_commentary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Commentary"
          }
        },
        "type": "object",
        "required": [
          "survey",
          "summary",
          "respondent_distribution",
          "questions"
        ],
        "title": "ReportResponse"
      },
...
      "Survey": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "objective": {
            "type": "string",
            "maxLength": 2000,
            "title": "Objective",
            "description": "조사 목적 (LLM 응답 품질 향상에 활용)",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "default": "draft"
          },
          "target_filter": {
            "$ref": "#/components/schemas/TargetFilter"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionConfig"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          },
          "persona_uuids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Persona Uuids",
            "description": "Step 2 확정 시점의 페르소나 스냅샷"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "target_filter",
          "execution",
          "created_at",
          "updated_at"
        ],
        "title": "Survey",
        "description": "설문 본체. 마법사 Step 1~4의 결과물."
      },
      "SurveyCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "objective": {
            "type": "string",
            "maxLength": 2000,
            "title": "Objective",
            "default": ""
          },
          "target_filter": {
            "$ref": "#/components/schemas/TargetFilter"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionConfig"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "title": "Questions"
          },
          "persona_uuids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Persona Uuids"
          }
        },
        "type": "object",
        "required": [
          "title",
          "target_filter",
          "execution"
        ],
        "title": "SurveyCreateRequest"
      },
      "SurveyListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "SurveyListResponse"
      },
...
      "TargetFilter": {
        "properties": {
          "age_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Min"
          },
          "age_max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Max"
          },
          "sex": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sex"
          },
          "provinces": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Provinces"
          },
          "family_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Family Types"
          },
          "education_levels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Education Levels"
          },
          "occupations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Occupations"
          },
          "query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "sampling": {
            "type": "string",
            "enum": [
              "all",
              "random_n",
              "proportional"
            ],
            "title": "Sampling",
            "default": "random_n"
          },
          "sample_size": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Sample Size",
            "default": 100
          }
        },
        "type": "object",
        "title": "TargetFilter",
        "description": "페르소나 선별 조건 — /api/dataset/personas/filter 입력과 호환."
      },
...
    }
  }
}